refactor: add a new package for ROS components (#26)
* refactor: modify process function returns processed image as optional


* refactor: add ros package


* fix: handle dot in fetch_parameters()


* refactor: use sensor_msgs::image_encodings instead of string


* refactor: use sensor_msgs::distortion_models instead of string


* refactor: modify the type of timestamp from uint64_t to int64_t


* fix: capture msg to prevent destructor is called before finishing process


* refactor: update default values


* refactor: add const qualifier


* refactor: enable to remap I/O topics in launcher


---------

+++ src/accelerated_image_processor_pipeline/test/test_utility.hpp
@@ -67,7 +67,7 @@ public:
   }
 
   const std::string frame_id = "camera";
-  const uint32_t timestamp = 123456789;
+  const int64_t timestamp = 123456789;
   const uint32_t width = 1920;
   const uint32_t height = 1080;
   const uint32_t step = width * 3;
